refactor: CouponManagement 컴포넌트 분리
